refactor: clean up repo root call
