+++ src/lib/config/cache.ts
@@ -12,17 +12,11 @@ let parentsRevList: revListT | undefined = undefined;
 
 let childrenRevList: revListT | undefined = undefined;
 
-let repoRootPath: string | undefined = undefined;
-
 let metaChildren: Record<string, Branch[]> | undefined = undefined;
 
 let branchList: Record<string, string[]> | undefined = undefined;
 
 class Cache {
-  public getRepoRootPath(): string | undefined {
-    return repoRootPath;
-  }
-
   public getBranchToRef(): Record<string, string> | undefined {
     return branchRefs?.branchToRef;
   }
@@ -69,10 +63,6 @@ class Cache {
 
   setBranchRefs(newBranchRefs: branchRefsT): void {
     branchRefs = newBranchRefs;
-  }
-
-  setRepoRootPath(newRepoRootPath: string): void {
-    repoRootPath = newRepoRootPath;
   }
 
   setMetaChildren(newMetaChildren: Record<string, Branch[]>): void {
