--- src/lib/config/repo_root_path.ts
@@ -1,17 +1,0 @@
-import { PreconditionsFailedError } from '../errors';
-import { gpExecSync } from '../utils/exec_sync';
-import { cache } from './cache';
-export function getRepoRootPath(): string {
-  const cachedRepoRootPath = cache.getRepoRootPath();
-  if (cachedRepoRootPath) {
-    return cachedRepoRootPath;
-  }
-  const repoRootPath = gpExecSync({
-    command: `git rev-parse --git-common-dir`,
-  });
-  if (!repoRootPath || repoRootPath.length === 0) {
-    throw new PreconditionsFailedError('No .git repository found.');
-  }
-  cache.setRepoRootPath(repoRootPath);
-  return repoRootPath;
-}
